agromate/clients/telegram-bot/requirements.txt
python-telegram-bot==13.15
fastapi>=0.97.0
uvicorn>=0.22.0
sqlalchemy>=2.0.15
alembic>=1.11.1
psycopg2-binary>=2.9.6
pydantic>=2.0.0
python-dotenv>=1.0.0
aiogram>=3.20.0
pydantic-settings>=2.8.1
httpx>=0.26.0
openai>=1.5.0
Pillow>=10.0.0
PyYAML>=6.0
aiohttp>=3.8.0
